Alert Types moved to /types.
