+++ src/pages/admin/index.tsx
@@ -23,7 +23,7 @@ import { useEffect } from 'react';
 import { useSelector } from 'react-redux';
 import { useHistory } from 'react-router-dom';
 import { RootState } from 'redux/store';
-import { UserState } from 'redux/typing';
+import { UserState } from 'types/redux/user';
 import { AdminPanelContainer } from './containers';
 
 const Admin = () => {
