+++ src/redux/typing.d.ts
@@ -18,16 +18,6 @@
 
 import type { Alert, AlertLevel, AlertText, AlertTimeout, AlertType } from "modules/alert/typing";
 
-
-export type AlertState = {
-    timeout?: AlertTimeout,
-    level?: AlertLevel,
-    type?: AlertType,
-    text?: AlertText
-};
-
-export type ShowAlertAction = Alert;
-
 export type SensorData = {
     epoch: number,
     value: number
@@ -63,51 +53,6 @@ export type InsertSensorsBulkDataAction = {
     [sensor: string]: Array<SensorData>
 };
 
-export type UpdateSensorsMetaAction = {
-    sensor: string,
-    key: string,
-    value: any
-};
-
-export type UserMeta = {
-    dept?: string,
-    lastLogin?: string,
-    createdAt?: string    
-};
-
-export type UserPrivilege = 'Anon' | 'Basic' | 'Admin' | 'Developer';
-
-export type UserDepartment = 'Electronics' | 'Operations' | 'Power-train' | 'Vehicle Performance' | 'Race Engineering' | 'Aerodynamics' | 'Tier 1' | 'NON SPECIFIED'
-
-export type SetUserAction = {
-    username: string,
-    accessToken: string,
-    creation: number,
-    privilege: UserPrivilege,
-    department: UserDepartment,
-    meta: UserMeta
-};
-
-export type RegisterUserAction = {
-    username: string, 
-    password: string, 
-    privilege: UserPrivilege, 
-    department: UserDepartment,
-}
-
-export type LoginUserAction = {
-    username: string,
-    password: string
-};
-
-export type UserState = {
-    username?: string,
-    accessToken?: string,
-    privilege?: UserPrivilege,
-    creation?: number,
-    department: UserDepartment,
-    meta: UserMeta
-};
 
 export type SessionState = {
     sessionName: string,
@@ -118,10 +63,3 @@ export type SessionState = {
     isRunning: boolean,
 }
 
-export type StartSessionAction = {
-    name: string,
-    driver: string, 
-    condition: string, 
-    sensors: string[],
-    groups: string[],
-}
